pyproject.toml
[project]
name = "rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "crewai[google-genai,tools]>=1.9.3",
    "faiss-cpu>=1.13.2",
    "litellm>=1.75.3",
    "ollama>=0.6.1",
    "pypdf>=6.7.0",
]
